Fixed moderniser
Here is a list of changes that were made to get this code to work.  Please not the comments, especially after we count the number of spaces in line 265 (something the original code is missing which is part of the reason it was not quite working).

Change line 122 so it returns [how_much, converted]  <switching it around causes a whole lot of pain>

In line 124, added (raw_unit) as a parameter
in line 125, added unit_tocheck = raw_unit

Added grams list to end of unit abbrievation list (line 139)
Added grams in unit_to_check (if / elif / else) block (lines 165 and 166)

Removed 'print(amount)' in line 240
Removed 'print(amount) in line 248 (in else statement after recipe line is split)

Added num_spaces = recipe_line.count(" ") to line 265 (ie: count how many spaces to work out if we have a number, unit and ingredient).  Then added if statement to line 266 and indented getting unit and ingredient (and checking unit with unit checker)

Indented everything after line 280 by one tab (up until output printed)

Edited everything in line 286 on-wards (see comments in code and compare with what you have).